+++ src/rx-database.ts
@@ -185,7 +185,9 @@ export class RxDatabaseBase<
         this.eventBulks$.next(changeEventBulk);
 
         // write to socket to inform other instances about the change
-        writeToSocket(this as any, changeEventBulk);
+        if (!this.storage.statics.doesBroadcastChangestream()) {
+            writeToSocket(this as any, changeEventBulk);
+        }
     }
 
     /**