+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 - Removed the `core` plugin. The default export `from 'rxdb'` now exports only the RxDB core without plugins.
 - Added optional `init()` function to `RxPlugin`.
+- Added `doesBroadcastChangestream()` to `RxStorageStatics`
+
 ### 11.6.0 (4 February 2022)
 
 Bugfixes: